pyproject.toml
[tool.poetry]
name = "cs-7643-efficiencylane"
version = "0.1.0"
description = ""
authors = ["datagero <datagero@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
adapters = "^0.1.2"
torch = "^2.2.2"
datasets = "^2.18.0"
accelerate = "^0.29.1"
transformers = "4.36.2"
hydra-core = "^1.3.2"
optuna = "^3.6.1"
mlflow = "^2.11.3"
optuna-dashboard = "^0.15.1"
scikit-learn = "<1.4.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
